separate callbacks from request processors
